# Improve contours approximations in ChessBoardDetector::generateQuads