@MaximSmolskiy 2024-03-04
> Does this PR replaces #24760?

Yes, I was going to close it after merging this PR