@AleksandrPanov 2024-03-06
I think that, before calling `approxPolyDP()`, we can check how close the contour is to a square/convex shape using `boundingRect()`/`minAreaRect()`/`convexHull()`. We can then drop some of the poor-quality contours at an earlier stage.

This is a suggestion for future improvements.